pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "threat-scanner"
version = "0.1.0"
description = "AI-powered supply chain security scanner for open source packages"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
threat-scan = "threat_scanner.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["e2e: end-to-end tests requiring a Lima VM (deselect with '-m not e2e')"]
filterwarnings = ["error"]

[tool.setuptools.packages.find]
where = ["src"]
